separate authService.js
